backend/requirements.txt
fastapi
uvicorn
websockets
ccxt
psycopg2
redis
pandas
xgboost
scikit-learn
joblib
sqlalchemy